Optimize packets by only care about the given world
Instead of checking against each player currently online on the server, we only care about the currently residing players for the given location's world. This not only drops the world name check, but also drasticly reduces the number of players to check if a world has almost no people in it, while the server might be packed.